+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog] and Pipelines adheres to
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
 ## [unreleased]
+### Change
+- Build with Composer 2.1; pin composer version to 2.1.11 (from 2.0.13),
+  since [0.0.55](#0055---2021-05-02)
 ### Fix
 - Container base directory creation with multiple directory components in
   `step.clone-path` (symlink errors)