+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog] and Pipelines adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## [unreleased]
+## [0.0.58](#0055---2021-11-07)
 ### Add
 - PHP 8.1 compatibility, build and docker images for pipelines.
 ### Change
@@ -601,4 +601,5 @@ The format is based on [Keep a Changelog] and Pipelines adheres to
 [0.0.55]: https://github.com/ktomk/pipelines/releases/tag/0.0.55
 [0.0.56]: https://github.com/ktomk/pipelines/releases/tag/0.0.56
 [0.0.57]: https://github.com/ktomk/pipelines/releases/tag/0.0.57
+[0.0.58]: https://github.com/ktomk/pipelines/releases/tag/0.0.58
 [unreleased]: https://github.com/ktomk/pipelines